Update on "add operation list for AutocastCPU"
[ghstack-poisoned]

+++ torch/cpu/amp/autocast_mode.py
@@ -5,5 +5,5 @@ class autocast(torch.autocast_mode.autocast):
     See :class:`torch.autocast`.
     ``torch.cpu.amp.autocast(args...)`` is equivalent to ``torch.autocast("cpu", args...)``
     """
-    def __init__(self, enabled=True, fast_dtype=torch.float16):
+    def __init__(self, enabled=True, fast_dtype=torch.bfloat16):
         super().__init__("cpu", enabled=enabled, fast_dtype=fast_dtype)
